% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions_binmatch.R
\name{bedtools_unionbedg}
\alias{bedtools_unionbedg}
\title{Run Bedtools unionbedg}
\usage{
bedtools_unionbedg(bgfiles, outfile)
}
\arguments{
\item{bgfiles}{Character list containing the list of BedGraphs to combine.
The names of the list elements will be used as output
header names. If no element names are set,
the file basename will be used instead.}
}
\description{
Run Bedtools unionbedg
}
\details{
Invoke System Command: \code{bedtools unionbedg}
}

