\name{CEMTscripts-package}
\alias{CEMTscripts-package}
\alias{CEMTscripts}
\docType{package}
\title{
\packageTitle{CEMTscripts}
}
\description{
\packageDescription{CEMTscripts}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CEMTscripts}
\packageIndices{CEMTscripts}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CEMTscripts}

Maintainer: \packageMaintainer{CEMTscripts}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
