% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromhmm_comparestates.R
\name{chromhmm_comparestates}
\alias{chromhmm_comparestates}
\title{Compute pearson correlation between the emission probabilities of different models}
\format{Numeric matrix created using \code{\%cor\%}}
\usage{
chromhmm_comparestates(model_file, reference_file)
}
\arguments{
\item{model_file}{ChromHMM model file}

\item{reference_file}{ChromHMM model file for comparison}
}
\description{
Compute pearson correlation between the emission probabilities of different models
}

