% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromhmm_comparestates.R
\name{chromhmm_comparestates_vizH}
\alias{chromhmm_comparestates_vizH}
\title{Compute pearson correlation between the emission probabilities of different models}
\usage{
chromhmm_comparestates_vizH(chromhmmcomparestates, ref = NULL, ...)
}
\arguments{
\item{chromhmmcomparestates}{Output of \code{chromhmm_comparestates()}.
Essentially a numeric matrix of values to be plotted.}

\item{ref}{Used to label the states of the reference model if the reference
model is Roadmap's 15 state core model "\code{roadmap5marks}"
or 18 state expanded model "\code{roadmap6marks}"}

\item{...}{Additional argument to \code{pheatmap(...)}}
}
\description{
Compute pearson correlation between the emission probabilities of different models
}
\examples{
library(dplyr)

## Specify model files
## Here we use Roadmap's 18 state model
## as both the model & reference files
model_file <-
  system.file("extdata", "model_18_core_K27ac.txt",
              package="CEMTscripts")
reference_file <- model_file

compstates <-
  chromhmm_comparestates(model_file, reference_file)

chromhmm_comparestates_vizH(compstates)
chromhmm_comparestates_vizH(
  compstates, ref="roadmap6marks", cluster_cols=F)
}

