% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromhmm_gencov.R
\name{chromhmm_gencov}
\alias{chromhmm_gencov}
\title{Load ChromHMM genomic coverage}
\usage{
chromhmm_gencov(enrichment_file, gencov_col = "Genome \%",
  isstateorder = FALSE)
}
\arguments{
\item{enrichment_file}{Enrichment overlap file containing genomic coverage.
The format of this file should be TSV, have headers,
and column 1 be the states.
When more than 1 file is given, the average is computed}

\item{gencov_col}{Name of genomic coverage column in the
\code{enrichment_file}.}

\item{isstateorder}{boolean; treat state as factor and order states?}
}
\description{
Load ChromHMM genomic coverage and find the average genomic coverage
if multiple enrichment files are given.
}

