% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromhmm_gencov.R
\name{chromhmm_getenrichmentcol}
\alias{chromhmm_getenrichmentcol}
\title{Pull out column(s) from ChromHMM overlap enrichment file}
\usage{
chromhmm_getenrichmentcol(enrichment_file, col)
}
\arguments{
\item{enrichment_file}{ChromHMM's enrichment overlap file}

\item{col}{Column(s) to pull out}
}
\description{
Pull out column(s) from ChromHMM overlap enrichment file
}

