% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromhmm_gencov.R
\name{chromhmm_loadenrichment}
\alias{chromhmm_loadenrichment}
\title{Load ChromHMM enrichment file}
\usage{
chromhmm_loadenrichment(enrichment_file)
}
\arguments{
\item{enrichment_file}{Assumes the first column of the TSV is chromatin state}
}
\description{
Load ChromHMM enrichment file
}

