% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromhmm_orderstates.R
\name{chromhmm_makestateorderingfile}
\alias{chromhmm_makestateorderingfile}
\title{Output state order as ChromHMM Reorder's stateorderingfile}
\usage{
chromhmm_makestateorderingfile(stateorder, filename = NULL)
}
\description{
Output state order as ChromHMM Reorder's stateorderingfile
}

