% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromhmm_orderstates.R
\name{chromhmm_orderstates}
\alias{chromhmm_orderstates}
\title{Crude ChomHMM state ordering
Crude ChromHMM state ordering according to Roadmap}
\usage{
chromhmm_orderstates(model)
}
\arguments{
\item{model}{output of \code{load_model()}}
}
\description{
Crude ChomHMM state ordering
Crude ChromHMM state ordering according to Roadmap
}

