% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromhmm_orderstates.R
\name{chromhmm_plt}
\alias{chromhmm_plt}
\title{Recreate ChromHMM emissions & transitions plot}
\usage{
chromhmm_plt(input, type, stateorder = NULL)
}
\arguments{
\item{type}{Specify what type of data the input is:
\code{e} for emissions or
\code{t} for transitions}

\item{stateorder}{character list specifying the state order}
}
\description{
Recreate ChromHMM emissions & transitions plot
}

