% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromhmm_pltenrichment.R
\name{chromhmm_pltenrichment}
\alias{chromhmm_pltenrichment}
\title{Plot overlap enrichment}
\usage{
chromhmm_pltenrichment(input, ...)
}
\arguments{
\item{input}{Output of \code{chromhmm_loadenrichment}}

\item{...}{arguments to pheatmap::pheatmap(...)}
}
\description{
Plot overlap enrichment
}

