% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromhmm_comparestates.R
\name{chromhmm_spreademissions}
\alias{chromhmm_spreademissions}
\title{Spread emissions}
\usage{
chromhmm_spreademissions(emissions)
}
\arguments{
\item{emissions}{dataframe produced by
\code{chromhmm_loadmodel(...)$emissions}}
}
\description{
Spread emissions
}

