% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combinematrix.R
\name{combine_matrix}
\alias{combine_matrix}
\title{Combine matrices}
\usage{
combine_matrix(matrix_list, method = "mean")
}
\arguments{
\item{matrix_list}{a list where each element is a matrix}

\item{method}{How should we combine the matrices?
One of "mean" (default),
standard deviation "sd", or
coefficient of variation "cov"}
}
\description{
Combine matrices
}

