% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_concordance.R
\name{create_freqmatrix}
\alias{create_freqmatrix}
\title{Create frequency matrix}
\usage{
create_freqmatrix(unionbedg, sample1, sample2)
}
\description{
To calculate the pairwise sample concordance,
we first create a frequency matrix
by comparing/counting the number of bins in each pairs of states
(one from sample1, the other from sample2).
}
\details{
After we have the frequency matrix,
we then use transition probabilities to account for transitions between states.
}

