% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_decisiveness.R
\name{decisiveness_average}
\alias{decisiveness_average}
\title{Calculate the average decisiveness across multiple states}
\usage{
decisiveness_average(emissions_probs)
}
\arguments{
\item{emissions_probs}{matrix containing the state emssions}
}
\description{
Calculate the average decisiveness across multiple states
}
\examples{
emissions_probs <- rbind(c(1, 1), c(1,0))
decisiveness_average(emissions_probs)
}

