% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_decisiveness.R
\name{decisiveness_model}
\alias{decisiveness_model}
\title{Calculate the decisiveness of model}
\usage{
decisiveness_model(emissions_probs, genomic_coverage = 1)
}
\arguments{
\item{emissions_probs}{matrix containing the state emssions}

\item{genomic_coverage}{list containing the genomic coverage of the states.
Used to scale the states.
By default, there is no scaling}
}
\description{
Calculate the decisiveness of model
}
\examples{
emissions_probs <- rbind(c(1, 0.5), c(1,0))
genomic_coverage <- c(.95, .05)
decisiveness_model(emissions_probs, genomic_coverage)
}

