% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemblgtf.R
\docType{data}
\name{ensemblgtf}
\alias{ensemblgtf}
\title{Protein coding transcripts}
\format{An object of class \code{GRanges} of length 90273.}
\source{
Downloaded from \url{http://www.ensembl.org/info/data/ftp/index.html}
}
\usage{
ensemblgtf
}
\description{
List of protein coding transcripts from Homo sapiens
obtained from Ensembl gene set annotations GR37.75
}
\examples{
## Object produce by the following:

# ensembl_gtf <- "Homo_sapiens.GRCh37.75.gtf"
ensemblgtf <- rtracklayer::import(ensembl_gtf)
ensemblgtf <-
  ensemblgtf[ensemblgtf$source=="protein_coding" &
               ensemblgtf$type=="transcript"]
save(ensemblgtf, file="data/ensemblgtf.RData")
}
\keyword{datasets}

