% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tallystate.R
\name{getfile_tallystate}
\alias{getfile_tallystate}
\title{Find tally state file in tally dir}
\usage{
getfile_tallystate(tally_dir, state = NULL)
}
\arguments{
\item{tally_dir}{directory containing the state files ending in "bed"}

\item{state}{Specify state file to be returned.
If \code{NULL} (default) is used,
all files bed files in tally_dir will be returned.}
}
\description{
Find tally state file in tally dir
}

