% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_createFlank.R
\name{gr_createBoundary}
\alias{gr_createBoundary}
\title{Create a list of boundary regions}
\usage{
gr_createBoundary(gr, width = 2000)
}
\arguments{
\item{gr}{gr object}

\item{width}{size of boundary region}
}
\description{
Create a list of boundary regions
}

