% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_createFlank.R
\name{gr_createFlank}
\alias{gr_createFlank}
\title{Create a list of flanking regions}
\usage{
gr_createFlank(gr, width = 2000)
}
\arguments{
\item{gr}{gr object}

\item{width}{size of flanking region}
}
\description{
Create a list of flanking regions
}

