% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromhmm_comparestates.R
\name{\%cor\%}
\alias{\%cor\%}
\title{Matrix pearson correlation}
\usage{
x \%cor\% y
}
\arguments{
\item{x}{\code{n * p} matrix}

\item{y}{\code{p * m} matrix}
}
\description{
Matrix pearson correlation
}
\details{
Computes the pearson correlation of
         the rows of x with the columns of y
}

