% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsynchronization_score.R
\name{load_emissionsdb}
\alias{load_emissionsdb}
\title{Load ChromHMM emission probabilities}
\usage{
load_emissionsdb(filename)
}
\arguments{
\item{filename}{The ChromHMM model emissions file}
}
\description{
Load ChromHMM emission probabilities
}

