% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_homer.R
\name{load_homer_knownresults}
\alias{load_homer_knownresults}
\title{Load HOMER knownResults.txt}
\format{Outputs a dataframe with the following columns:
        \code{tf, dnabindingdomain, origin, source, consensus,
        pval, pvalLOG, qvalBEN, tagetN, tagetP, bgN, bgP}}
\source{
\url{http://homer.salk.edu/homer/motif/motifDatabase.html}
}
\usage{
load_homer_knownresults(knownresults_file, returncounts = FALSE)
}
\arguments{
\item{knownresults_file}{path to the "knownResults.txt" file}

\item{returncounts}{boolean; If TRUE, will return
vector of target & background sequences}
}
\description{
Load HOMER knownResults.txt
}

