% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_macs2.R
\name{load_macs2_broadpeak}
\alias{load_macs2_broadpeak}
\title{Load MACS2 NAME_peaks.broadPeak}
\source{
\url{https://github.com/taoliu/MACS}
}
\usage{
load_macs2_broadpeak(broadPeak_file)
}
\description{
Load MACS2 NAME_peaks.broadPeak
}

