% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_macs2.R
\name{load_macs2_narrowpeak}
\alias{load_macs2_narrowpeak}
\title{Load MACS2 NAME_peaks.narrowPeak}
\source{
\url{https://github.com/taoliu/MACS}
}
\usage{
load_macs2_narrowpeak(narrowPeak_file)
}
\description{
Load MACS2 NAME_peaks.narrowPeak
}

