% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tallystate.R
\name{load_tallystate}
\alias{load_tallystate}
\title{Load unionbedg tally state file}
\usage{
load_tallystate(input_file)
}
\arguments{
\item{input_file}{TSV file with 6 columns + no header:
chrom, start, stop,
number of samples, sample representation
e.g. 0010 (where 0=absent; 1=present),
number of bins in region i.e. (stop-start)/200bp.}
}
\description{
Load unionbedg tally state file
}

