% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_concordance.R
\name{load_transitiondb}
\alias{load_transitiondb}
\title{Load ChromHMM transition probabilities}
\usage{
load_transitiondb(filename)
}
\arguments{
\item{filename}{The ChromHMM model transitions file}
}
\description{
Load ChromHMM transition probabilities
}

