% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_concordance.R
\name{load_unionbedg}
\alias{load_unionbedg}
\title{Load unionbedg}
\usage{
load_unionbedg(filename, binsize = 200)
}
\arguments{
\item{filename}{TSV file containing the unionbedg of ChromHMM sample segments.
The file contains the following columns:
\code{chrom}, \code{start}, \code{end}, <sampleID ...>;
Note: header is included in the file.}

\item{binsize}{The integer bin size used in ChromHMM (default = 200).}
}
\description{
Load unionbedg
}
\details{
\code{filename} is created using bedtools \code{unionbedg}.
}

