% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_concordance.R
\name{make_sample_pairs}
\alias{make_sample_pairs}
\title{Make sample pairs}
\usage{
make_sample_pairs(the_samples)
}
\arguments{
\item{the_samples}{character list of sample ids}
}
\description{
From a character list of samples,
create a dataframe specifying the pairs of samples
to compare for sample concordance.
}
\examples{
the_samples <- c("CEMT_33", "CEMT_34", "CEMT_50")
(sample_pairs <- make_sample_pairs(the_samples))
}

