% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeExpression_fromLabelcol.R
\name{makeExpression_fromLabelcol}
\alias{makeExpression_fromLabelcol}
\title{make Expression counts for UpSetR plots}
\usage{
makeExpression_fromLabelcol(input, isgr = TRUE, column.group = "label",
  column.tally = NULL, group.names = NULL)
}
\arguments{
\item{input}{either a dataframe or gr object}

\item{isgr}{boolean; is \code{input} a gr object?
(i.e. output of \code{regions_get()} or
\code{regions_addgenes()})}

\item{column.group}{Name of column in mcols(gr) to group by;
should contain only "0" and "1",
length be equal to number of samples,
"0" = sample is absent and
"1" = sample is present.}

\item{column.tally}{Name of column in mcols(gr) to add up
when we group by \code{column.group}.
Default NULL, we count each row in gr as 1.}

\item{group.name}{character vector of column.group sample names}
}
\description{
Count groups for for use in
UpSetR::fromExpression(.) %>% UpSetR::upset()
}

