% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromhmm_orderstates.R
\name{markstate_order}
\alias{markstate_order}
\title{Order states}
\usage{
markstate_order(emissions, mark, threshold = 0.05)
}
\description{
For a given mark and emission probabilties above a threshold,
order the states according to decreasing emission probabilties.
}

