% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combinematrix.R
\name{plot_combinematrix}
\alias{plot_combinematrix}
\title{Combine a list of matrices and plot heatmap}
\usage{
plot_combinematrix(mat, is_list = TRUE, method = "mean",
  use_manual_opt = FALSE, ...)
}
\arguments{
\item{mat}{either a list where each element is a matrix,
or a matrix object}

\item{is_list}{boolean; is \code{mat} a list?}

\item{method}{when \code{mat} is a list of matrices,
how should we combine the matrices?
"mean" (default), "sd", or "cov" (coefficient of variation)}

\item{use_manual_opt}{boolean; choose your own pheatmap parameters?}

\item{...}{additional arguments to pheatmap::pheatmap}
}
\description{
Combine a list of matrices and plot heatmap
}

