% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sets.R
\name{plot_sets}
\alias{plot_sets}
\title{Plot sets}
\usage{
plot_sets(x, type = "upset", ...)
}
\arguments{
\item{x}{a list whereby each element is a character vector/set and
the name of the element is the name of the set}

\item{type}{what kind of plot to produce: one of "upset" (default)
or "venn"}

\item{...}{additional arguments to \code{UpSetR::upset(...)} or
\code{VennDiagram::venn.diagram(...)}}
}
\description{
Plot sets
}
\examples{
x <-
  list(
    set1 = letters[1:10],
    set2 = c("a", "e", "i", "o", "u"),
    set3 = letters[1:3]
  )
plot_sets(x)
plot_sets(x, type="venn")
}

