% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions_binmatch.R
\name{regions_binmatch}
\alias{regions_binmatch}
\title{State comparison}
\format{List object with 3 names
  \itemize{
    \item{\code{raw}}{
       Dataframe produced with bedtools unionbedg}
    \item{\code{summary_viz}}{
       ggplot summarizing the number of bins in the states of interest
       in the two samples}
    \item{\code{summary}}{
       Dataframe containing the number of bins which have matching states}
  }}
\usage{
regions_binmatch(filenames, state)
}
\arguments{
\item{filenames}{Vector containing the paths of 2 ChromHMM segmentations.
Element names will be used as the id.}

\item{state}{List specifying the states in a genomic feature
to be treated the same.}
}
\description{
Do the particular states of a genomic feature
between the segmentation of a sample match?
}
\details{
Intended to be used to compare the states between the same bins
  of the same sample generated using different models
}

