% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions_get.R
\name{regions_get}
\alias{regions_get}
\title{Get regions}
\usage{
regions_get(tally_dir, state, nsamples = NULL, regex.in = "*",
  regex.notin = NULL)
}
\arguments{
\item{tally_dir}{directory containing the state files ending in "bed"}

\item{state}{integer; state file in \code{tally_dir} to load}

\item{nsamples}{Filter: How many samples must region overlap?
Can be single element e.g. "4" or
a numeric list e.g. "c(3, 4)".
Specifying \code{NULL} (default) will keep everything.}

\item{regex.in}{Filter: Which samples must be in the result?
Character length must equal to the number of samples.
Only "0", "1", and "." characters are permitted, where
"1" = present, "0" = absent, "." = either present or absent.
e.g. ".1001" means in the 5 samples:
sample 1 can be present or absent;
samples 2 and 5 must be present; and
samples 3 and 4 must be absent in the results.}

\item{regex.notin}{Filter: Which samples must not be in the result?
Requirements are the same as \code{regex.in}}
}
\description{
Get the list of regions overlaping n samples in a particular state
}

