% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions_togenes.R
\name{regions_togenes}
\alias{regions_togenes}
\title{Convert regions to genes}
\format{Dataframe with 4 columns: "gene_id", "samples_maxoverlap",
  "samples_neighbor", and "label_neighbor".
  samples_maxoverlap refers to the maximum number of samples
  overlapping the same bin of a gene; samples_neighbor refers
  to the samples in the total number of unique samples in the
  gene neighborhood, identity of which is given in label_neighborhood.}
\usage{
regions_togenes(regionsaddgenes)
}
\arguments{
\item{regionsaddgenes}{Output of \code{regions_addgenes()}.
Assumes metadata contain columns:
"samples" (number of overlapping samples) and "gene_id"}
}
\description{
Convert regions to genes
}

