% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_concordance.R
\name{sample_concordance}
\alias{sample_concordance}
\title{Calculate sample concordance}
\usage{
sample_concordance(the_samples, filename.transitions, filename.unionbedg,
  verbose = FALSE)
}
\arguments{
\item{the_samples}{character list of sample ids used for the pairwise comparisons}

\item{filename.transitions}{The ChromHMM model transitions file}

\item{filename.unionbedg}{TSV file containing the unionbedg of ChromHMM sample segments.
The file contains the following columns:
\code{chrom}, \code{start}, \code{end}, <sampleID ...>;
Note: header is included in the file.}

\item{verbose}{boolean; print messages?}
}
\description{
Calculate sample concordance
}
\details{
The output will be a dataframe containing 4 columns:
         \code{s1, s2, dist_max, dist_avg}. s1 and s2 refers to
         all pairwise combinations of \code{the_samples}.
         dist_max refers to the distance score using 1-max transition prob;
         dist_avg refers to the distance score using 1-mean transition prob.
}

