% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_concordance.R
\name{sample_concordance_summary}
\alias{sample_concordance_summary}
\title{Sample concordance summary}
\usage{
sample_concordance_summary(sampleconcordance)
}
\arguments{
\item{sampleconcordance}{the output of \code{sample_concordance()}}
}
\description{
Computes the mean, median, min, max, sd
of the sample pairs from \code{sample_concordance()}
}

