% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_concordance.R
\name{sample_concordance_viz}
\alias{sample_concordance_viz}
\title{Sample concordance visualization}
\usage{
sample_concordance_viz(sampleconcordance, type = "max")
}
\arguments{
\item{sampleconcordance}{the output of \code{sample_concordance()}}
}
\description{
Sample concordance visualization
}

