% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_concordance.R
\name{sample_concordance_vizH}
\alias{sample_concordance_vizH}
\title{Sample concordance visualization (heatmap)}
\usage{
sample_concordance_vizH(sampleconcordance, type = "max",
  fun_scale = function(x) {     1 - (x/15181508) }, ...)
}
\arguments{
\item{sampleconcordance}{the output of \code{sample_concordance()}}

\item{type}{One of "\code{max}" (default) or "\code{avg}";
see \code{state_distance()} for details.
This parameter chooses which distance metric to use.}

\item{fun_scale}{function to scale matrix;
use "\code{NULL}" for no scaling
(i.e use the raw sample distances.
See details.}

\item{...}{additional arguments to \code{pheatmap(...)}}
}
\description{
Visualize sample concordance in terms of a heatmap
}
\details{
The value 15181508 is the number of 200 bp bins as segmented by ChromHMM.
  x/15181508 accounts for the total distance (i.e. error) between samples.
  1-(x/15181508) transforms the distance to a similarity measure.
  For scaling the distance between 0 and 1, use
  \code{function(x){(x-min(x,na.rm=T)) / (max(x,na.rm=T)-min(x,na.rm=T))}}
}

