% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_concordance.R
\name{sample_distance}
\alias{sample_distance}
\title{Distance between 2 states}
\usage{
sample_distance(freqmatrix, transition_probs, type = "max")
}
\description{
Get the distance score from the transition probabilities between 2 samples
}

