% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep_colbychar.R
\name{sep_colbychar}
\alias{sep_colbychar}
\title{Separate column by each character}
\usage{
sep_colbychar(dat, column = "label", column.names = NULL, keep = TRUE)
}
\arguments{
\item{dat}{dataframe}

\item{column}{The name of the column you want to split.
This column should be character; if not,
this column will be converted to character.}

\item{column.names}{The new column names after the split}

\item{keep}{boolean; keep original column in dataframe?}
}
\description{
Separate column by each character
}
\examples{
dat <- head(iris)
sep_colbychar(dat, column="Petal.Length")
}

