% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sets.R
\name{set_total}
\alias{set_total}
\title{Total unique elements in embedded list}
\usage{
set_total(x, getunique = FALSE)
}
\arguments{
\item{x}{a list whereby each element is a character vector/set and
the name of the element is the name of the set}

\item{getunique}{boolean; if TRUE, will instead return the
list of unique elements}
}
\description{
Total unique elements in embedded list
}

