% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_decisiveness.R
\name{state_decisiveness}
\alias{state_decisiveness}
\title{Calculate state decisiveness}
\usage{
state_decisiveness(filename.emissions)
}
\arguments{
\item{filename.emissions}{The ChromHMM model emissions file}
}
\description{
Calculate, based on state emissions, the average decisiveness of a model
}

