% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_concordance.R
\name{state_distance}
\alias{state_distance}
\title{Distance between 2 states}
\usage{
state_distance(transition_probs, x, y, type = "max")
}
\arguments{
\item{transition_probs}{output of \code{load_transitiondb(...)}}

\item{x}{state from sample1}

\item{y}{state from sample2}

\item{type}{Specify which probability to use:
"\code{max}" (default) or "\code{avg}".
Specifying a type not in the list will give you "max".
See details.}
}
\description{
Get the distance score from the transition probabilities between 2 states
}
\details{
Let A and B be the states of sample1 and sample2 in a given bin.
Let P(.) be the transition probability of a state.
When A==B, the score returned is 0.
When type=="avg", the score returned is 1-mean(P(A->B), P(B->A)).
When type=="max", the score returned is 1-max(P(A->B), P(B->A)).
}

