% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statecolors.R
\docType{data}
\name{statecolors}
\alias{statecolors}
\title{State colors}
\format{List of 12 colors}
\usage{
statecolors
}
\description{
State colors
}
\details{
Produced by: \code{
c(
  "tssactive" = "#FF0000",
  "tss" = "#FF4500",
  "txn" = "#008000",
  "txnweak" = "#006400",
  "znf" = "#66CDAA",
  "enhgenic" = "#C2E105",
  "enhactive" = "#FFC34D",
  "enh" = "#FFFF00",
  "bivtss" = "#CD5C5C",
  "bivenh" = "#BDB76B",
  "heterochromatin" = "#8A91D0",
  "repeats" = "#C0C0C0",
  "polycomb" = "#808080",
  "low" = "#FFFFFF"
)
save(statecolors, file = "data/statecolors.rda")
}
}
\keyword{datasets}

