% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statecolors.R
\name{statecolors_makepalette}
\alias{statecolors_makepalette}
\title{Make statepalette}
\usage{
statecolors_makepalette(x)
}
\arguments{
\item{x}{List of states. Valid states are described in \code{?statecolors}}
}
\description{
Make statepalette
}
\details{
For use as \code{statepalette} in \code{tally_state_summary_viz()}
}

