% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ta_auc.R
\name{ta_auc}
\alias{ta_auc}
\title{Compute middle Riemann sum}
\usage{
ta_auc(transcriptabundance, ngenes = 10000)
}
\arguments{
\item{transcriptabundance}{output of \code{transcript_abundance}}

\item{ngenes}{integer; how many top expr genes?}
}
\description{
Compute middle Riemann sum
}

