% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_abundance.R
\name{ta_getgenes}
\alias{ta_getgenes}
\title{Return the list of top genes}
\usage{
ta_getgenes(transcriptabundance, ngenes)
}
\arguments{
\item{transcriptabundance}{output of \code{transcript_abundance}}

\item{ngenes}{integer; how many top expr genes?}
}
\description{
Return the list of top genes
}

