% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_abundance.R
\name{ta_ngenes2proportion}
\alias{ta_ngenes2proportion}
\title{Proportion of expressed transcripts that make up the top ngenes}
\usage{
ta_ngenes2proportion(transcriptabundance, ngenes, getgenes = FALSE)
}
\arguments{
\item{transcriptabundance}{output of \code{transcript_abundance}}

\item{ngenes}{integer; how many top expr genes?}

\item{getgenes}{boolean; if TRUE, will instead return the list of top ngenes}
}
\description{
Proportion of expressed transcripts that make up the top ngenes
}

