% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_abundance.R
\name{ta_proportion2ngenes}
\alias{ta_proportion2ngenes}
\title{Number of genes which make up a proportion of expressed transcript}
\usage{
ta_proportion2ngenes(transcriptabundance, p, gt = FALSE, getgenes = FALSE)
}
\arguments{
\item{transcriptabundance}{output of \code{transcript_abundance}}

\item{p}{proportion. Value should be between 0 - 1.}

\item{gt}{boolean. If FALSE (default), the function will return the
number of genes equal to or below the specified proportion.
If TRUE, the function will return +1 gene which makes
the proportion slightly larger than the specified value}

\item{getgenes}{boolean; if TRUE, will instead return the list of top ngenes}
}
\description{
Number of genes which make up a proportion of expressed transcript
}

