% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_state_summary.R
\name{tally_state}
\alias{tally_state}
\title{Summarize overlaps in a given state}
\usage{
tally_state(input_file)
}
\arguments{
\item{input_file}{TSV file with 6 columns + no header:
chrom, start, stop,
number of samples, sample representation
e.g. 0010 (where 0=absent; 1=present),
number of bins in region i.e. (stop-start)/200bp.}
}
\description{
In a given state, tally the number of bins which overlap
}

