% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_state_consensusprop.R
\name{tally_state_consensusprob}
\alias{tally_state_consensusprob}
\title{Consensus proportion}
\format{Dataframe with 4 columns sorted by increasing "proportion" size
  \itemize{
   \item{\code{state}}
   \item{\code{bins_consensus}}
   \item{\code{bins_atleast1}}
   \item{\code{proportion}}
  }}
\usage{
tally_state_consensusprob(tallystatesummary)
}
\arguments{
\item{tallystatesummary}{Output of \code{tally_state_summary()}}
}
\description{
Proportion of bins marked-in-all-samples among bins marked-in-at-least-1-sample
}
\details{
The number of bins marked as state \code{i} in the consensus divided by
  the number of bins marked as state \code{i} in at least 1 sample
  (...which gives somewhat a measure of variability).
}

