% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_state_summary.R
\name{tally_state_summary}
\alias{tally_state_summary}
\title{Summarize overlaps for all states}
\usage{
tally_state_summary(tally_dir, verbose = FALSE, cache = TRUE)
}
\arguments{
\item{tally_dir}{directory containing the state files}

\item{verbose}{boolean; print messages?}

\item{cache}{boolean; If TRUE, results will be saved to
\code{tally_dir}/tallystate_summary.txt &/or
data will be loaded from this file}
}
\description{
In a given state, tally the number of bins which overlap.
Assumes input files end in "bed" and
a state e.g. state 1's filename is named "*.U1.*", "*.E1.*" *.1.*"
}

