% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_state_summary_viz.R
\name{tally_state_summary_viz}
\alias{tally_state_summary_viz}
\title{Visualize overlaps for all states}
\usage{
tally_state_summary_viz(input, statepalette = NULL, facet_wrap = FALSE,
  scale_y_log10 = FALSE, shape = FALSE, plottheme = NULL)
}
\arguments{
\item{input}{output of \code{tally_state_summary()}
i.e. dataframe where column 1 is state, and the remaining
column are the number of bins which overlap 1, 2, ... samples.}

\item{statepalette}{character list where each element is a color and
the name of each element is a \code{state} level.}

\item{facet_wrap}{boolean}

\item{scale_y_log10}{boolean; when this is TRUE, the bin counts will be
log10 transformed and the plot will use points instead of bars.}

\item{shape}{boolean; when this is TRUE, the bin counts are
rescaled to range 0-1 and the
\code{scale_y_log10} argument is not used}

\item{plottheme}{argument to adjust ggplot theme}
}
\description{
Plot output of \code{tally_state_summary()} into a barchart
}

