% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_abundance.R
\name{transcript_abundance}
\alias{transcript_abundance}
\title{Transcript abundance}
\format{Dataframe whose rows are sorted in descending order based on \code{expr_column}.
        There are 5 columns: gene_id, expr, cumsum, proportion, and count}
\usage{
transcript_abundance(txi_df, expr_column)
}
\arguments{
\item{txi_df}{one element (i.e. "abundance") from \code{transcript_togene}}

\item{expr_column}{column containing transcript information}
}
\description{
Transcript abundance
}

