% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_abundance.R
\name{transcript_abundance_viz}
\alias{transcript_abundance_viz}
\title{Plot Transcript abundance}
\usage{
transcript_abundance_viz(transcriptabundance, ngenes = 10000)
}
\arguments{
\item{transcriptabundance}{output of \code{transcript_abundance}}

\item{ngenes}{integer; how many top expr genes to plot?}
}
\description{
Use a dot plot to plot the transcript abundance
}

