% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_abundance.R
\name{transcript_cumsum}
\alias{transcript_cumsum}
\title{Cumulative sum of a (desc) sorted vector}
\usage{
transcript_cumsum(x)
}
\arguments{
\item{x}{vector of elements}
}
\description{
Cumulative sum of a (desc) sorted vector
}
\examples{
x <- c("a"=1, "b"=3, "c"=NA, "d"=6, "e"=2, "f"=NA)
CEMTscripts:::transcript_cumsum(x)
}

