% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_togene.R
\name{transcript_togene}
\alias{transcript_togene}
\title{Summarize genes}
\usage{
transcript_togene(files, file_ids = NULL)
}
\arguments{
\item{files}{a character vector of filenames
for the transcript-level abundances}

\item{file_ids}{a character vector of ids in the same order as \code{files}.
By default, the directory name of the base file is used.}
}
\description{
Uses tximport to map Salmon transcripts to genes.
Column names/sample ids are added at this step.
}
\details{
There are 5 columns in the Salmon quant file: name, length,
  effective length, TPM, and number of reads.

  tximport outputs 3 matrices:
  \itemize{
    \item{\code{abundance}}{
    sum of transcript TPMs for a given gene}
    \item{\code{counts}}{
    sum of transcript read counts for a given gene}
    \item{\code{length}}{
    TPM weighted average of transcripts for a given gene}
  }
}

