% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_togene.R
\name{transcript_tx2gene}
\alias{transcript_tx2gene}
\title{Create tx2gene for use in tximport}
\format{a two-column data.frame linking
        transcript id (column 1) to gene id (column 2)}
\source{
Ensembl BioMart download of "Homo sapiens genes (GRCh37.p13)"
        with attributes: "Ensembl Gene ID" and "Ensembl Transcript ID"
        \url{http://grch37.ensembl.org/biomart}
}
\usage{
transcript_tx2gene()
}
\description{
Create tx2gene for use in tximport
}

