% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_togene.R
\name{transcript_tximport}
\alias{transcript_tximport}
\title{Map transcripts to genes}
\usage{
transcript_tximport(files, ...)
}
\arguments{
\item{files}{a character vector of filenames
for the transcript-level abundances}
}
\description{
Uses tximport to map Salmon transcripts to genes
}

