% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsynchronization_score.R
\name{unsync}
\alias{unsync}
\title{Calculate the unsynchronization of 1 state}
\usage{
unsync(state_emissions)
}
\arguments{
\item{state_emissions}{list of state emissions from histone modifications}
}
\description{
Calculate the unsynchronization of 1 state
}
\details{
Calculated by \code{d_k = sum[h=1 to H]{min(1-E_h, E_h)}}
         want to minimize this value.
}

