% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsynchronization_score.R
\name{unsynchronization}
\alias{unsynchronization}
\title{Calculate the unsynchronization of the model}
\usage{
unsynchronization(emissions_probs, scale = 1, returntotal = TRUE)
}
\arguments{
\item{emissions_probs}{matrix containing the state emssions}

\item{scale}{list used to scale the states;
by default, there is no scaling.}

\item{returntotal}{boolean; return the total value (TRUE) or
the list of values making up the total (FALSE)}
}
\description{
Calculate the unsynchronization of the model
}
\details{
Calculated by \code{D = sum[k=1 to K]{d_k * s_k}};
         want to minimize this value. s_k represent the scaled term.
}
\examples{
emissions_probs <- rbind(c(1, 0.5), c(1,0))
scale <- c(.95, .05)
CEMTscripts:::unsynchronization(emissions_probs, scale)
}

