% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsynchronization_score.R
\name{unsynchronization_score}
\alias{unsynchronization_score}
\title{Calculate model unsynchronization from chromhmm emissions file}
\usage{
unsynchronization_score(filename.emissions, returntotal = TRUE)
}
\arguments{
\item{filename.emissions}{The ChromHMM model emissions file}

\item{returntotal}{boolean; return the total value (TRUE) or
the list of values making up the total (FALSE)}
}
\description{
Calculate, based on state emissions, the unsynchronization of a model
}
\details{
Calculated by \code{D = sum[k=1 to K]{d_k * s_k}} where
         \code{d_k = sum[h=1 to H]{min(1-E_h, E_h)}} and s_k is the
         scale term. Note: want to choose a model which minimizes the
         unsynchronization (D) term.
}

