% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_state_summary.R
\name{visualize_state_summary}
\alias{visualize_state_summary}
\title{Visualize overlaps for all states}
\usage{
visualize_state_summary(input, statepalette = NULL, facet_wrap = FALSE)
}
\arguments{
\item{input}{output of \code{tally_state_summary()}}

\item{statepalette}{character list where each element is a color and
the name of each element is the state number}

\item{facet_wrap}{boolean}
}
\description{
Plot output of \code{tally_state_summary()} into a barchart
}

