% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_decisiveness.R
\name{decisiveness}
\alias{decisiveness}
\title{Calculate the decisiveness of 1 state}
\usage{
decisiveness(state_emissions)
}
\arguments{
\item{state_emissions}{list of state emissions from histone modifications}
}
\description{
Calculate the decisiveness of 1 state
}
\details{
Calculated by \code{d_k = sum[h=1 to H]{min(1-E_h, E_h)}}
         want to minimize this value.
}

