% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_colorstrip.R
\name{plot_colorstrip}
\alias{plot_colorstrip}
\title{Function to add color to facet strip}
\usage{
plot_colorstrip(d, p, statepalette)
}
\arguments{
\item{d}{the data}

\item{p}{the base plot}

\item{statepalette}{character list where each element is a color and
the name of each element is the state number}
}
\description{
Function to add color to facet strip
}

