% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions_togenes.R
\name{regions_addgenes}
\alias{regions_addgenes}
\title{Label regions with gene ids}
\format{Format is the same as \code{regions}, but with an extra "gene_id" column.
  The length of the output depends on the number of
  matching regions with gene_ids.}
\usage{
regions_addgenes(regions, regiontype = "tss", tss.upstream = 2000,
  tss.downstream = 2000)
}
\arguments{
\item{regions}{GRanges object}

\item{tss.upstream}{When \code{regiontype} is "tss", this argument is used
to specify the upstream GenomicRanges::promoter size}

\item{tss.downstream}{See \code{tss.upstream}}
}
\description{
Label regions with gene ids
}

