% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_decisiveness.R
\name{state_decisiveness}
\alias{state_decisiveness}
\title{Calculate state decisiveness}
\usage{
state_decisiveness(filename.emissions, genomic_coverage, returntype = "")
}
\arguments{
\item{filename.emissions}{The ChromHMM model emissions file}

\item{genomic_coverage}{list containing the genomic coverage of the states.
Used to scale the states.}

\item{returntype}{specify what value is returned:
"d_k" for vector of d_k;
"dc_k" for scaled vector of d_k * genomic coverage;
anything else (default) for the metric itself}
}
\description{
Calculate, based on state emissions, the average decisiveness of a model
}
\details{
Calculated by \code{D = sum[k=1 to K]{d_k * c_k}} where
         \code{d_k = sum[h=1 to H]{min(1-E_h, E_h)}};
         want to minimize this value.
}

