% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_state_summary.R
\name{tally_state}
\alias{tally_state}
\title{Summarize overlaps in a given state}
\usage{
tally_state(input_file)
}
\arguments{
\item{input_file}{}
}
\description{
In a given state, tally the number of bins which overlap
}

