% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_state_summary.R
\name{tally_state_summary}
\alias{tally_state_summary}
\title{Summarize overlaps for all states}
\usage{
tally_state_summary(tally_dir, out_file = NULL)
}
\arguments{
\item{tally_dir}{directory containing the state files}
}
\description{
In a given state, tally the number of bins which overlap.
Assumes input files end in "bed" and
a state i.e. state 1 is named ".U1." or ".1."
}

