% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_state_summary_viz.R
\name{tally_state_summary_viz}
\alias{tally_state_summary_viz}
\title{Visualize overlaps for all states}
\usage{
tally_state_summary_viz(input, statepalette = NULL, facet_wrap = FALSE)
}
\arguments{
\item{input}{output of \code{tally_state_summary()}
i.e. dataframe where column 1 is state, and the remaining
column are the number of bins which overlap 1, 2, ... samples.}

\item{statepalette}{character list where each element is a color and
the name of each element is a \code{state} level.}

\item{facet_wrap}{boolean}
}
\description{
Plot output of \code{tally_state_summary()} into a barchart
}

